+++ CHANGELOG.md
@@ -4,6 +4,27 @@
 
 From version 2.0.0 onward, release notes are published on the [GitHub Releases page](https://github.com/intent-hq/cloudlands-releases/releases). Auto-generated entries for 2.x releases also appear below.
 
+## [2.122.1](https://github.com/intent-hq/cloudlands-fe/compare/v2.122.0...v2.122.1) (2026-09-01)
+
+
+### 🐛 Bug Fixes
+
+* prevent effect_update_depth_exceeded on dedicated-agent route ([#1970](https://github.com/intent-hq/cloudlands-fe/issues/1970)) ([97e7f6a](https://github.com/intent-hq/cloudlands-fe/commit/97e7f6a6daf66a0dce22843c99c414dc22dc00a7))
+
+## [2.122.0](https://github.com/intent-hq/cloudlands-fe/compare/v2.121.0...v2.122.0) (2026-08-31)
+
+
+### 🚀 Features
+
+* refine project sidebar workspace actions ([#1966](https://github.com/intent-hq/cloudlands-fe/issues/1966)) ([5191ced](https://github.com/intent-hq/cloudlands-fe/commit/5191ced77ae1e1e76bdc84271f352bd4d0fece5b))
+
+## [2.121.0](https://github.com/intent-hq/cloudlands-fe/compare/v2.120.0...v2.121.0) (2026-08-31)
+
+
+### 🚀 Features
+
+* surface agent file locks in the Changes panel ([#1959](https://github.com/intent-hq/cloudlands-fe/issues/1959)) ([a9f9dd7](https://github.com/intent-hq/cloudlands-fe/commit/a9f9dd73424c6bafa54353068e94a36e90426986))
+
 ## [2.120.0](https://github.com/intent-hq/cloudlands-fe/compare/v2.119.0...v2.120.0) (2026-08-31)